Squashed commit of the following:

commit 45bc7d00bcde706a4df09ec5ba416892a79f7b09
Author: Victor Roemer <viroemer@cisco.com>
Date:   Fri Jul 29 00:06:55 2016 -0400

    Convert sd_pattern to Hyperscan engine.

+++ src/ips_options/sd_credit_card.cc
@@ -61,7 +61,7 @@ static inline int CheckIssuers(const uint8_t *cardnum, uint32_t buflen)
  *
  * Returns: 1 on match, 0 otherwise.
  */
-int SdLuhnAlgorithm(const uint8_t *buf, uint32_t buflen)
+int SdLuhnAlgorithm(const uint8_t *buf, unsigned long long buflen)
 {
     int i, digits, alternate, sum, val;
     char cc_digits[CC_COPY_BUF_LEN]; /* Normalized CC# string */
